refactor: Use enum in Panel, move function
